refactor routes and split app sections
