+++ src/layout/RoutingErrorPage.css
@@ -1,4 +1,4 @@
-.ErrorPage {
+.RoutingErrorPage {
   height: 100vh;
   display: flex;
   flex-direction: column;
